Rearranging the namespace.
