+++ panoptes/utils/data/__init__.py
@@ -1,0 +1,1 @@
+from .assets import Downloader
